chore(operators): add days-to-month exercise

+++ fundamentals/operators/final-grade.js
@@ -2,9 +2,9 @@ const prompt = require("prompt-sync")()
 
 const TOTAL_GRADE = 3
 const MAX_DIGITS = 2
-const grade1 = parseInt(prompt("Nota #1 "))
-const grade2 = parseInt(prompt("Nota #2 "))
-const grade3 = parseInt(prompt("Nota #3 "))
+const grade1 = parseFloat(prompt("Nota #1 "))
+const grade2 = parseFloat(prompt("Nota #2 "))
+const grade3 = parseFloat(prompt("Nota #3 "))
 
 const average = (grade1 + grade2 + grade3) / TOTAL_GRADE
